Introduce tracing and improve the collision detection
Use the tracing* crates under the "debug" feature flag to allow for logging into a the file (tracing.log).
Introduce also the --manual-ball cmd flag to disable automatic ball movements and control it using a keyboard.
Make sure that if the ball collides with an object it's shifted to only the object's outer bounds.